refactor: Split up analysis handler
- move different handlers into separate files

Closes: #165
